Refactor: Users could not work with unsigned messages
Solution: Refactor the project, rename Message as SignedMessage and add
a new Message model that does not contain signature related fields.
