+++ aleph_message/__init__.py
@@ -1,1 +1,11 @@
-from .models import Message, MessagesResponse
+from aleph_message.message import (Message,
+                                   PostMessage, AggregateMessage,
+                                   StoreMessage, ForgetMessage, ProgramMessage)
+
+from aleph_message.signed import (SignedMessage,
+                                  SignedPostMessage,
+                                  SignedAggregateMessage,
+                                  SignedStoreMessage,
+                                  SignedForgetMessage,
+                                  SignedProgramMessage,
+                                  SignedMessagesResponse)
